Revert "Refactor the ssh client code"
